# openapi.yaml
openapi: 3.0.4
info:
  title: SOA lab API
  version: 0.0.1
components:
  schemas:
    Color:
      type: string
      enum:
        - RED
        - BLUE
        - ORANGE
        - WHITE
        - BROWN
    Country:
      type: string
      enum:
        - SPAIN
        - VATICAN
        - THAILAND
...
    Error:
      type: object
      properties:
        message:
          type: string
        code:
          type: integer
  parameters:
...
    Nationality:
      name: nationality
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/Country'
    EyeColor:
      name: eye-color
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/Color'
...
paths:
...
  /demography/nationality/{nationality}/hair-color:
    get:
      description: "Get all people with this nationality and this hair-color"
      parameters:
        - $ref: '#/components/parameters/Nationality'
      responses:
        '200':
          description: "Count retrieved successfully"
          content:
            application/xml:
              schema:
                type: object
                properties:
                  count:
                    type: integer
              examples:
                countExample:
                  summary: "Example of count result"
                  value: |
                    <count>3</count>
        '500':
          description: "Internal server error"
          content:
            application/xml:
              schema:
                $ref: '#/components/schemas/Error'
  /demography/nationality/{nationality}/eye-color/{eye-color}:
    get:
      description: "Get all people with this nationality and this eye-color"
      parameters:
        - $ref: '#/components/parameters/Nationality'
        - $ref: '#/components/parameters/EyeColor'
      responses:
        '200':
          description: "Count retrieved successfully"
          content:
            application/xml:
              schema:
                type: object
                properties:
                  count:
                    type: integer
              examples:
                eyeColorCountExample:
                  summary: "Example of eye color count result"
                  value: |
                    <count>2</count>
        '500':
          description: "Internal server error"
          content:
            application/xml:
              schema:
                $ref: '#/components/schemas/Error'